requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
WeasyPrint==59.0
Jinja2==3.1.2
SQLAlchemy==2.0.21
WTForms==3.0.1
pandas==2.0.3
openpyxl==3.1.2
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow==10.0.0
reportlab==4.0.4
redis==5.0.1
pytest==7.4.0
pytest-flask==1.2.0
pandas==2.0.3
openpyxl==3.1.2